sync: split SyncOrchestrator.kt and dedup step-failure handling
SyncOrchestrator.kt held the orchestrator plus the sync state types and
the Hilt entry point. Move SyncState/ItemProgress/SyncStep/SyncSummary to
SyncState.kt (next to their owner, SyncProgressReporter) and the entry
point to SyncOrchestratorEntryPoint.kt. Same package, so no import
churn; finalizeSyncResult stays with the orchestrator it finalizes.

Replace the six copies of `finishError(error, false); return Error(error)`
in syncAllNotebooks with a private failStep() helper. Behavior-preserving:
only the preflight/folder steps that already called finishError are
converted; the later steps' handling is unchanged.

Addresses P16 readability (1f).
